backend/requirements.txt
# Use the small CPU version of Torch to stay under the 512MB RAM limit of Free Tiers
--extra-index-url https://download.pytorch.org/whl/cpu
torch
torchvision
ultralytics
fastapi
uvicorn
python-multipart
numpy
opencv-python-headless
pillow
scipy
pandas
matplotlib
lxml
roboflow
motor
pymongo
apscheduler
passlib[bcrypt]
PyJWT
